+++ CHANGELOG.md
@@ -2,8 +2,20 @@
 
 All notable changes to this project will be documented in this file.
 
-<!-- git-cliff-unreleased-start -->
-## 3.1.1 - **not yet released**
+## [3.1.2](https://github.com/apify/apify-client-python/releases/tag/v3.1.2) (2026-08-10)
+
+### 🐛 Bug Fixes
+
+- Read file-like KVS values before upload and reject unencodable ones ([#965](https://github.com/apify/apify-client-python/pull/965)) ([064b5cc](https://github.com/apify/apify-client-python/commit/064b5cccbce8b585f58271b089e4987c6559a7d8)) by [@vdusek](https://github.com/vdusek)
+- Respect caller-supplied Content-Encoding for pre-compressed request bodies ([#997](https://github.com/apify/apify-client-python/pull/997)) ([b5541c9](https://github.com/apify/apify-client-python/commit/b5541c919923f69d537b4062e21c1f3c46396c7a)) by [@vdusek](https://github.com/vdusek), closes [#996](https://github.com/apify/apify-client-python/issues/996)
+
+### ⚡ Performance
+
+- Skip request-body compression for already-compressed content types ([#987](https://github.com/apify/apify-client-python/pull/987)) ([808bdde](https://github.com/apify/apify-client-python/commit/808bdde06ed1befe9d30bede0f947320240747a6)) by [@vdusek](https://github.com/vdusek)
+- Skip request-body compression for small payloads ([#988](https://github.com/apify/apify-client-python/pull/988)) ([6bd31b2](https://github.com/apify/apify-client-python/commit/6bd31b2abb840ea9af1efc98bf984c85b8510ce3)) by [@vdusek](https://github.com/vdusek), closes [#934](https://github.com/apify/apify-client-python/issues/934)
+
+
+## [3.1.1](https://github.com/apify/apify-client-python/releases/tag/v3.1.1) (2026-08-03)
 
 ### 🐛 Bug Fixes
 
@@ -12,7 +24,6 @@ All notable changes to this project will be documented in this file.
 - Keep pagination iterators advancing past fully-filtered pages ([#964](https://github.com/apify/apify-client-python/pull/964)) ([2e7f75c](https://github.com/apify/apify-client-python/commit/2e7f75ca951f2b7b59c319badd6e024d47fe67f1)) by [@vdusek](https://github.com/vdusek)
 
 
-<!-- git-cliff-unreleased-end -->
 ## [3.1.0](https://github.com/apify/apify-client-python/releases/tag/v3.1.0) (2026-07-20)
 
 ### 🚀 Features